+++ src/collidermodus.cc
@@ -37,7 +37,13 @@ ColliderModus::ColliderModus(Configuration modus_config,
   if (modus_cfg.has_value({"Calculation_Frame"})) {
     frame_ = modus_cfg.take({"Calculation_Frame"});
   }
-
+  if (fermi_motion_ == FermiMotion::On) {
+    logg[LCollider].info() << "Fermi motion is ON.";
+  } else if (fermi_motion_ == FermiMotion::Frozen) {
+    logg[LCollider].info() << "FROZEN Fermi motion is on.";
+  } else if (fermi_motion_ == FermiMotion::Off) {
+    logg[LCollider].info() << "Fermi motion is OFF.";
+  }
   Configuration proj_cfg = modus_cfg.extract_sub_configuration({"Projectile"});
   Configuration targ_cfg = modus_cfg.extract_sub_configuration({"Target"});
   /* Needed to check if projectile and target in customnucleus are read from
@@ -51,6 +57,11 @@ ColliderModus::ColliderModus(Configuration modus_config,
     same_file = same_inputfile(proj_cfg, targ_cfg);
     projectile_ = std::make_unique<CustomNucleus>(
         proj_cfg, params.testparticles, same_file);
+  } else if (proj_cfg.has_value({"Alpha_Clustered"})) {
+    logg[LCollider].info() << "Projectile is alpha-clustered with woods-saxon "
+                              "parameters for the He-clusters listed below.";
+    projectile_ = create_alphaclustered_nucleus(proj_cfg, params.testparticles,
+                                                "projectile");
   } else {
     projectile_ = std::make_unique<Nucleus>(proj_cfg, params.testparticles);
   }
@@ -65,6 +76,11 @@ ColliderModus::ColliderModus(Configuration modus_config,
   } else if (targ_cfg.has_value({"Custom"})) {
     target_ = std::make_unique<CustomNucleus>(targ_cfg, params.testparticles,
                                               same_file);
+  } else if (targ_cfg.has_value({"Alpha_Clustered"})) {
+    logg[LCollider].info() << "Target is alpha-clustered with woods-saxon "
+                              "parameters for the He-clusters listed below.";
+    target_ =
+        create_alphaclustered_nucleus(targ_cfg, params.testparticles, "target");
   } else {
     target_ = std::make_unique<Nucleus>(targ_cfg, params.testparticles);
   }
@@ -199,7 +215,7 @@ ColliderModus::ColliderModus(Configuration modus_config,
         "Please provide one of Sqrtsnn/E_Kin/P_Lab.");
   }
   if (energy_input > 1) {
-    throw std::domain_error(
+    throw std::invalid_argument(
         "Input Error: Redundant collision energy. "
         "Please provide only one of Sqrtsnn/E_Kin/P_Lab.");
   }
@@ -217,7 +233,7 @@ ColliderModus::ColliderModus(Configuration modus_config,
       if (sampling_ == Sampling::Custom) {
         if (!(modus_cfg.has_value({"Impact", "Values"}) ||
               modus_cfg.has_value({"Impact", "Yields"}))) {
-          throw std::domain_error(
+          throw std::invalid_argument(
               "Input Error: Need impact parameter spectrum for custom "
               "sampling. "
               "Please provide Values and Yields.");
@@ -226,7 +242,7 @@ ColliderModus::ColliderModus(Configuration modus_config,
             modus_cfg.take({"Impact", "Values"});
         const std::vector<double> yields = modus_cfg.take({"Impact", "Yields"});
         if (impacts.size() != yields.size()) {
-          throw std::domain_error(
+          throw std::invalid_argument(
               "Input Error: Need as many impact parameter values as yields. "
               "Please make sure that Values and Yields have the same length.");
         }
@@ -260,14 +276,6 @@ ColliderModus::ColliderModus(Configuration modus_config,
     // the displacement is half the distance (both nuclei are shifted
     // initial_z_displacement_ away from origin)
     initial_z_displacement_ /= 2.0;
-  }
-
-  if (fermi_motion_ == FermiMotion::On) {
-    logg[LCollider].info() << "Fermi motion is ON.";
-  } else if (fermi_motion_ == FermiMotion::Frozen) {
-    logg[LCollider].info() << "FROZEN Fermi motion is on.";
-  } else if (fermi_motion_ == FermiMotion::Off) {
-    logg[LCollider].info() << "Fermi motion is OFF.";
   }
 
   FluidizationType IC_type = modus_cfg.take({"Initial_Conditions", "Type"},
@@ -315,17 +323,17 @@ std::unique_ptr<DeformedNucleus> ColliderModus::create_deformed_nucleus(
       !nucleus_cfg.has_value({"Deformed", "Beta_2"});
 
   if (automatic_deformation && was_any_deformation_parameter_given) {
-    throw std::domain_error(
+    throw std::invalid_argument(
         "Automatic deformation of " + nucleus_type +
         " nucleus requested, but deformation parameter(s) were provided as"
         " well. Please, check the 'Deformed' section in your input file.");
   } else if (!automatic_deformation && !was_any_beta_given) {
-    throw std::domain_error(
+    throw std::invalid_argument(
         "Manual deformation of " + nucleus_type +
         " nucleus requested, but no deformation beta parameter was provided."
         " Please, check the 'Deformed' section in your input file.");
   } else if (!automatic_deformation && was_gamma_given_without_beta_2) {
-    throw std::domain_error(
+    throw std::invalid_argument(
         "Manual deformation of " + nucleus_type +
         " nucleus requested, but 'Gamma' parameter was provided without "
         "providing a value of 'Beta_2' having hence no deformation effect. "
@@ -333,6 +341,32 @@ std::unique_ptr<DeformedNucleus> ColliderModus::create_deformed_nucleus(
   } else {
     return std::make_unique<DeformedNucleus>(nucleus_cfg, ntest,
                                              automatic_deformation);
+  }
+}
+
+std::unique_ptr<AlphaClusteredNucleus>
+ColliderModus::create_alphaclustered_nucleus(Configuration &nucleus_cfg,
+                                             int ntest,
+                                             const std::string &nucleus_type) {
+  bool automatic_alphaclustering =
+      nucleus_cfg.take({"Alpha_Clustered", "Automatic"});
+  bool was_sidelength_given =
+      nucleus_cfg.has_value({"Alpha_Clustered", "Sidelength"});
+
+  if (automatic_alphaclustering && was_sidelength_given) {
+    throw std::invalid_argument(
+        "Automatic alpha-clustering of " + nucleus_type +
+        " nucleus requested, but a sidelength was provided as"
+        " well. Please, check the 'Alpha_Clustered' section in your input "
+        "file.");
+  } else if (!automatic_alphaclustering && !was_sidelength_given) {
+    throw std::invalid_argument(
+        "Manual alpha-clustering of " + nucleus_type +
+        " nucleus requested, but no sidelength was provided."
+        " Please, check the 'Alpha_Clustered' section in your input file.");
+  } else {
+    return std::make_unique<AlphaClusteredNucleus>(nucleus_cfg, ntest,
+                                                   automatic_alphaclustering);
   }
 }
 
@@ -373,7 +407,7 @@ double ColliderModus::initial_conditions(Particles *particles,
     target_->generate_fermi_momenta();
   } else if (fermi_motion_ == FermiMotion::Off) {
   } else {
-    throw std::domain_error("Invalid Fermi_Motion input.");
+    throw std::invalid_argument("Invalid Fermi_Motion input.");
   }
 
   // Boost the nuclei to the appropriate velocity.
@@ -469,7 +503,7 @@ std::pair<double, double> ColliderModus::get_velocities(double s, double m_a,
       v_a = fixed_target_projectile_v(s, m_a, m_b);
       break;
     default:
-      throw std::domain_error(
+      throw std::invalid_argument(
           "Invalid reference frame in "
           "ColliderModus::get_velocities.");
   }