+++ src/include/smash/collidermodus.h
@@ -12,6 +12,7 @@
 #include <string>
 #include <utility>
 
+#include "alphaclusterednucleus.h"
 #include "deformednucleus.h"
 #include "forwarddeclarations.h"
 #include "fourvector.h"
@@ -215,6 +216,18 @@ class ColliderModus : public ModusDefault {
    * \return Pointer to the created deformed nucleus object.
    * **/
   static std::unique_ptr<DeformedNucleus> create_deformed_nucleus(
+      Configuration &nucleus_cfg, const int ntest,
+      const std::string &nucleus_type);
+  /** Configure Alpha-Clustered Nucleus
+   *
+   * Sets up an alpha-clustered nucleus object based on the input parameters in
+   * the configuration file. \param[in] nucleus_cfg Subset of configuration,
+   * projectile or target section. \param[in] ntest Number of test particles
+   * \param[in] nucleus_type String 'projectile' or 'target'. To display an
+   * appropriate error message.
+   * \return Pointer to the created deformed nucleus object.
+   * **/
+  static std::unique_ptr<AlphaClusteredNucleus> create_alphaclustered_nucleus(
       Configuration &nucleus_cfg, const int ntest,
       const std::string &nucleus_type);
   /**